Changed a few things after reading the "conventions for devices APIs" document
* Created an enum for the registers
* Changed all floats to doubles
* Changed a few properties and methods from private to protected
* Changed the returned temperature value to Temperature struct

+++ src/devices/Ad7193/AdcValue.cs
@@ -4,7 +4,7 @@
     {
         public byte Channel { get; set; }
         public uint Raw { get; set; }
-        public float Voltage { get; set; }
+        public double Voltage { get; set; }
         public long Time { get; set; }
 
         public override string ToString()
